refactor: make `BallotImage` builder an `impl` method
